Addons are now loaded dynamically.
Created a helper class to deal with loading / registering addons to the
game. In the play state we now call game.addons.loadEnabledAddons() to
fire them up.  This lays the ground work for beeing able to
enable/disable addons in the menu as you please. Before I go any further
and start working on cleaning up the actual addon code I would like some
feedback on this commit, just to make sure it makes sense & we're all on
the same page.